+++ tests/Unit/CollectionNormalizerTest.php
@@ -7,9 +7,9 @@ use Symfony\Component\PropertyInfo\Extractor\ReflectionExtractor;
 use Symfony\Component\Serializer\Encoder\JsonEncoder;
 use Symfony\Component\Serializer\Normalizer\PropertyNormalizer;
 use Symfony\Component\Serializer\Serializer as SymfonySerializer;
-use Thunk\Verbs\Lifecycle\StateManager;
 use Thunk\Verbs\SerializedByVerbs;
 use Thunk\Verbs\State;
+use Thunk\Verbs\State\StateManager;
 use Thunk\Verbs\Support\Normalization\CarbonNormalizer;
 use Thunk\Verbs\Support\Normalization\CollectionNormalizer;
 use Thunk\Verbs\Support\Normalization\NormalizeToPropertiesAndClassName;
